requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
numpy==1.24.3
Pillow==10.0.1
deep-sort-realtime==1.3.2
filterpy==1.4.5
scipy==1.11.4
alembic==1.12.1
python-dotenv==1.0.0
